requirements.txt
streamlit==1.26.0
snowflake-snowpark-python==1.22.1
snowflake.core==0.9.0
trulens-connectors-snowflake
trulens-core
trulens-providers-cortex
trulens-dashboard